NEW VERSION/requirements.txt
# ============================================================================
# JARVIS AI Assistant - Complete Requirements
# Python Version: 3.10.8
# Generated: January 24, 2026
# ============================================================================

# GUI Framework
PyQt6==6.7.0
PyQt6-Qt6==6.7.3
PyQt6_sip==13.11.0

# Core System Libraries
colorama==0.4.6
psutil==7.1.3
python-dotenv==1.2.1
pywin32==311
pypiwin32==223
comtypes==1.4.15

# Web & Network
requests==2.32.5
urllib3==2.6.3
certifi==2026.1.4
websockets==15.0.1
websocket-client==1.9.0
Flask==3.1.2
Werkzeug==3.1.5
itsdangerous==2.2.0
blinker==1.9.0
h11==0.16.0

# Computer Vision & Image Processing
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
mediapipe==0.10.14
pillow==12.1.0

# GUI Automation
pyautogui==0.9.54
PyAutoGUI==0.9.54
MouseInfo==0.1.3
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pyperclip==1.11.0

# Audio Control & Processing
pycaw==20251023
vosk==0.3.45
sounddevice==0.5.3
soundfile==0.13.1
pyaudio==0.2.14
audioread==3.1.0
soxr==1.0.0

# Speech & TTS
pyttsx3==2.90
SpeechRecognition==3.10.4
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_en==2.0.1

# Machine Learning & AI (CPU optimized)
torch==2.3.0+cpu
torchvision==0.18.0+cpu
torchaudio==2.3.0+cpu
transformers==4.45.0
sentencepiece==0.2.0
tokenizers==0.20.3
huggingface-hub==0.36.0
safetensors==0.7.0

# Scientific Computing
numpy==1.26.4
scipy==1.15.3
pandas==2.3.3
scikit-learn==1.7.2
numba==0.63.1
llvmlite==0.46.0

# Audio & Signal Processing
librosa==0.11.0
srt==3.5.3
joblib==1.5.3
lazy_loader==0.4
msgpack==1.1.2
pooch==1.8.2
soundfile==0.13.1

# Deep Learning Support
coqpit==0.0.17
einops==0.8.1
jax==0.6.2
jaxlib==0.6.2
opt_einsum==3.4.0
absl-py==2.3.1
flatbuffers==25.12.19
protobuf==4.25.8
ml_dtypes==0.5.4

# Intel Math Kernel Library
intel-openmp==2021.4.0
mkl==2021.4.0
tbb==2021.13.1

# NLP & Text Processing
inflect==7.5.0
num2words==0.5.14
python-crfsuite==0.9.12
regex==2025.10.23

# Web Scraping & Automation
selenium==4.25.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
soupsieve==2.8.3
trio==0.32.0
trio-websocket==0.12.2
wsproto==1.3.2
sniffio==1.3.1
sortedcontainers==2.4.0
outcome==1.3.0.post0
PySocks==1.7.1

# Communication & Messaging
pywhatkit==5.4
mtranslate==1.8

# Network & Speed Testing
speedtest-cli==2.1.3

# Location & Geography
geopy==2.4.1
geographiclib==2.1

# Date & Time Handling
python-dateutil==2.9.0
pytz==2025.2
tzdata==2025.3
tzlocal==5.3.1
dateparser==1.1.8
babel==2.17.0

# Data Visualization & Plotting
matplotlib==3.10.8
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9

# Utilities & Helpers
tqdm==4.67.1
click==8.3.1
filelock==3.20.3
fsspec==2026.1.0
packaging==26.0
platformdirs==4.5.1
typing_extensions==4.15.0
attrs==25.4.0
decorator==5.2.1
more-itertools==10.8.0
six==1.17.0
threadpoolctl==3.6.0
docopt==0.6.2
exceptiongroup==1.3.1
jsonlines==1.2.0
networkx==3.4.2
sympy==1.14.0
mpmath==1.3.0
typeguard==4.4.4

# Wikipedia API
wikipedia==1.4.0

# Parsing & Configuration
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
pyparsing==3.3.2
pycparser==3.0
cffi==2.0.0

# Character Encoding
charset-normalizer==3.4.4
idna==3.11
